:bug: Fixed RT indicator
